requirements.txt
deepgram-sdk
python-dotenv
textblob
scikit-learn
pandas
numpy
streamlit
transformers
torch
